change int to long in repository

+++ src/main/java/com/laura/carpaciu/dao/interfaces/CompanyRepository.java
@@ -11,7 +11,7 @@ import com.laura.carpaciu.entity.clients.Company;
 //This will be AUTO IMPLEMENTED by Spring into a Bean called companyRepository
 //CRUD refers Create, Read, Update, Delete
 
-public interface CompanyRepository extends CrudRepository<Company, Integer>  {
+public interface CompanyRepository extends CrudRepository<Company, Long>  {
 
 	Optional<Company> findByCui(String cui);
