+++ src/main/java/com/laura/carpaciu/dao/interfaces/InvoiceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.laura.carpaciu.entity.invoice.Invoice;
 import com.laura.carpaciu.entity.order.ServiceOrder;
 
-public interface InvoiceRepository extends CrudRepository<Invoice, Integer>{
+public interface InvoiceRepository extends CrudRepository<Invoice, Long>{
 
 	void saveInvoiceToDatabase(Invoice invoice);
