+++ src/main/java/com/laura/carpaciu/dao/interfaces/LuminaireCasesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.laura.carpaciu.entity.order.LuminaireCases;
 
-public interface LuminaireCasesRepository extends CrudRepository<LuminaireCases, Integer> {
+public interface LuminaireCasesRepository extends CrudRepository<LuminaireCases, Long> {
 
 //	void createLuminaireCases(LuminaireCases luminaireProblems);
 //
