+++ src/main/java/com/laura/carpaciu/dao/interfaces/LuminaireRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.laura.carpaciu.entity.clients.Company;
 import com.laura.carpaciu.entity.luminaire.Luminaire;
 
-public interface LuminaireRepository extends CrudRepository<Luminaire, Integer> {
+public interface LuminaireRepository extends CrudRepository<Luminaire, Long> {
 	Optional<Luminaire> findBySerialNumber(String serialNumber);
 
 	void create(Luminaire luminaire);
