+++ src/main/java/com/laura/carpaciu/dao/interfaces/PieceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.laura.carpaciu.entity.clients.Company;
 import com.laura.carpaciu.entity.luminaire.Piece;
 
-public interface PieceRepository extends CrudRepository<Piece, Integer> {
+public interface PieceRepository extends CrudRepository<Piece, Long> {
 
 	Optional<Piece> findByPartNumber(String partNumber);
