+++ src/main/java/com/laura/carpaciu/dao/interfaces/WorkPriceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.laura.carpaciu.entity.clients.Company;
 import com.laura.carpaciu.entity.work.WorkPrice;
 
-public interface WorkPriceRepository extends CrudRepository<WorkPrice, Integer> {
+public interface WorkPriceRepository extends CrudRepository<WorkPrice, Long> {
 
 	Optional<WorkPrice> findAllWorkPrices();
